refactor(create-sanity): extract flag parsing to parseArgs.ts
Move all parseArgs logic (option building, alias resolution, negation
merging, option/exclusive validation, help printing) out of index.ts
into a dedicated parseArgs.ts module.
